requirements.txt
streamlit
opencv-python
pyzbar
numpy
Pillow
